misc changes flush

+++ bootseq/loader/shared/safeparsemaxint.c
@@ -8,11 +8,11 @@
 #include "fsys.h"
 
 int
-safe_parse_maxint (char **str_ptr, int *myint_ptr)
+safe_parse_maxint (char **str_ptr, long *myint_ptr)
 {
   char *ptr = *str_ptr;
-  int myint = 0;
-  int mult = 10, found = 0;
+  long myint = 0;
+  long mult = 10, found = 0;
 
   /*
    *  Is this a hex number?
